# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Phantasmagoria API",
    "description": "API documentation for the Alumni Influencers AR platform. Built to serve data to client-agnostic AR applications while managing a blind bidding influencer marketplace.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token here to access authenticated Alumni routes."
      },
      "apiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Enter the Developer API Key to access the AR Client endpoints."
      }
    },
    "schemas": {
      "BidPayload": {
        "type": "object",
        "required": ["bidAmount", "targetDate"],
        "properties": {
          "bidAmount": {
            "type": "number",
            "format": "float",
            "description": "The monetary amount of the bid. Must be greater than 0 and higher than the user's previous bid.",
            "example": 50.50
          },
          "targetDate": {
            "type": "string",
            "format": "date",
            "description": "The future date the user is bidding to be Alumni of the Day.",
            "example": "2026-08-01"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/profile/": {
      "get": {
        "summary": "Get full alumni profile and associated records",
        "tags": ["Profile Management"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Returns full profile object including degrees, certifications, etc." },
          "401": { "description": "Unauthorized" },
          "404": { "description": "Profile not found." }
        }
      }
    },
...
    "/api/bids": {
      "post": {
        "summary": "Place or update a blind bid",
        "description": "Places a new bid or increases an existing bid for a specific date. Enforces the 3-win monthly limit and the increase-only rule.",
        "tags": ["Bidding"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BidPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bid placed or updated successfully"
          },
          "400": {
            "description": "Bad Request - Invalid amount, past date, or bid not increased"
          },
          "403": {
            "description": "Forbidden - Monthly 3-win limit reached"
          }
        }
      }
    },
...
  }
}